moved get_all_files to utils.py
